api_server/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
deepface
tensorflow
ulid-py
psycopg2-binary
python-dotenv